requirements.txt
google-cloud-texttospeech==0.1.0
boto3
awscli
gTTS
